pyproject.toml
[tool.poetry]
name = "spn_eval"
version = "0.1.0"
description = ""
authors = ["Yuri Baburov <burchik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
jupyter = "^1.0"
notebook = "^6.0"
google-cloud-speech = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
